Collection Page Changes
Simplified the Collections page and worked on client.js to connect to ME api better
