+++ public/client.js
@@ -1,3 +1,25 @@
+const express = require("express");
+
+const app = express();
+const port = process.env.SERVER_PORT || 8000;
+
+// Add Access Control Allow Origin headers
+app.use((req, res, next) => {
+  res.setHeader("Access-Control-Allow-Origin", "*");
+  res.header(
+    "Access-Control-Allow-Headers",
+    "Origin, X-Requested-With, Content-Type, Accept"
+  );
+  next();
+});
+
+app.get("/api", (req, res) => {
+  res.json("Hello");
+});
+
+app.listen(port, () => console.log(`Listening on port ${port}`));
+
+
 // client.js
 const options = { method: 'GET', headers: { accept: 'application/json' } };
